flatten gfx
