+++ gfx/src/console.rs
