+++ gfx/src/torus/mesh.rs
